requirements.txt
streamlit
opencv-python
pillow
requests
toml